backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
